Move YearInput classes to yearinput package.
