+++ docs/com/checkout/android_sdk/Input/YearInput.html
@@ -117,7 +117,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.widget.AppCompatSpinner</li>
 <li>
 <ul class="inheritance">
-<li>com.checkout.sdk.input.YearInput</li>
+<li>com.checkout.sdk.yearinput.YearInput</li>
 </ul>
 </li>
 </ul>
@@ -371,7 +371,7 @@ extends android.support.v7.widget.AppCompatSpinner</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a name="setYearListener-com.checkout.sdk.input.YearInput.YearListener-">
+<a name="setYearListener-com.checkout.sdk.yearinput.YearInput.YearListener-">
 <!--   -->
 </a>
 <ul class="blockListLast">
